\name{SGPstateData}
\alias{SGPstateData}
\docType{data}
\title{State assessment program data from large scale state assessments for use with SGP package}
\description{
An environment (an object of class environment)  containing information on state assessment programs, organized by state. Currently the environment contains achievement level cutscores and 
labels for the state assessments, assessment name and abbreviation, growth cutscores and labels, information on vertical scaling, and knots and boundaries currently being used for SGP analyses
at the state level. The cutscores, in particular, are used to calculate growth-to-standard/projection values. States currently included in the data set are Arizona (AZ), California (CA), 
Colorado (CO), Connecticut (CT), Georgia (GA), Hawaii (HI), Idaho (ID), Indiana (IN), Kansas (KS), Maine (ME), Massachusetts (MA), Minnesota (MN), Mississippi (MS),
Missouri (MO), Nevada (NV), New Hampshire (NH), New Jersey (NJ), New York (NY), Oregon (OR), Rhode Island (RI), Utah (UT), Vermont (VT), Virginia (VA), West Virginia (WV), 
Wisconsin (WI), Archdioces of Baltimore (AOB), Colorado English Language Assessment (CELA), Demonstration (DEMO), Australia NAPLAN (NAPLAN) and New Haven (NEW_HAVEN).  
}
\usage{data(SGPstateData)}

\author{Damian W. Betebenner \email{dbetebenner@nciea.org} and Adam Van Iwaarden \email{vaniwaarden@colorado.edu}}

\source{State assessment data and technical assessment documentation}
\keyword{datasets}
