% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slmw_mcmc.R
\name{slmw_mcmc}
\alias{slmw_mcmc}
\title{Spatial Linear Regression Modeling Weibull Distributions}
\usage{
slmw_mcmc = function(nsim, W, S, Sstar, startingPoint = NULL,
                      details = FALSE, forecasting = TRUE, burn = 0,
                      thin = 1, verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{W}{List of sublists containing Weibull realizations.}

\item{S}{Matrix of covariates for observed point.}

\item{Sstar}{Matrix of covariates for interpolating points.}

\item{startingPoint}{List for starting point.}

\item{details}{If it is true then also values of Gamma, Delta, and Pi are
collected (large memory consumption).}

\item{forecasting}{If it true then hyper-parameters values at interpolating
points are collected.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
A named list containing a posterior sample.
}
\description{
Markov chain Monte Carlo (MCMC) algorithm for linear spatial model.
}
\details{
MCMC sampler for linear spatial model without the binomial random
variables. Details are the same of \code{slmbw_mcmc} except the part for
the binomial distributions.
}
\references{
...
}
